add fact to actor:

+++ Source/Game/Engine/Engine/Engine.h
@@ -34,7 +34,7 @@ struct Engine
     bool running() const { return m_running;};
     void shutdown() { m_running = false;};
 
-    void apply_framelimit(double timeMS);
+    void apply_framelimit(uint32_t timeMS);
     void core_init();
     void subsystem_init();
     void core_shutdown();
